+++ CHANGELOG.md
@@ -8,6 +8,12 @@ have notable changes.
 
 Changes since v2.38
 
+### Changes
+- Create license admin page:
+  - When creating attachment license, the save button no longer activates until all attachments have been successfully uploaded. (#3292)
+  - Attachment upload error is indicated by small icon next to the upload button (detailed error is still shown in top of the page).
+  - Pending attachment upload is indicated by spinner next to the upload button.
+
 ### Fixes
 - Resolved issue where workflow voting could not be removed, which caused UI to display raw translation keys due to nil voting values. (#3357)